+++ tests/transformers/models/causal_lm_models/test_blocking.py
@@ -1,0 +1,180 @@
+# -----------------------------------------------------------------------------
+#
+# Copyright (c) Qualcomm Technologies, Inc. and/or its subsidiaries.
+# SPDX-License-Identifier: BSD-3-Clause
+#
+# -----------------------------------------------------------------------------
+
+
+import pytest
+
+from .check_causal_models import check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100
+from .config import (
+    QEFF_TEST_PROFILE,
+    causal_lm_models_dict,
+    compile_params,
+    export_params,
+    generate_params,
+    test_models_causal,
+    transform_params,
+)
+
+
+@pytest.mark.llm
+@pytest.mark.non_qaic
+@pytest.mark.parametrize("model_name", test_models_causal)
+def test_fp16_export_compile_blocking_CB(model_name):
+    """
+    Fp16 + Subfunction + CB + Blocking enabled
+    """
+    if causal_lm_models_dict.get(model_name, None) == model_name and QEFF_TEST_PROFILE == "tiny_model":
+        pytest.skip("Skipping it is not a tiny model and will run in nightly tests.")
+
+    HEAD_BLOCK_SIZE = 8
+    NUM_KV_BLOCKS = 2
+    NUM_Q_BLOCKS = 2
+
+    # head blocking only
+    qaic_config = dict(enable_blocking=True, head_block_size=HEAD_BLOCK_SIZE)
+    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
+        model_name,
+        continuous_batching=True,
+        transform_params={**transform_params, "qaic_config": qaic_config},
+        export_params=export_params,
+        compile_params=compile_params,
+        generate_params=generate_params,
+        export_compile_only=True,
+    )
+
+    # kv blocking only
+    qaic_config = dict(enable_blocking=True, num_kv_blocks=NUM_KV_BLOCKS)
+    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
+        model_name,
+        continuous_batching=True,
+        transform_params={**transform_params, "qaic_config": qaic_config},
+        export_params=export_params,
+        compile_params=compile_params,
+        generate_params=generate_params,
+        export_compile_only=True,
+    )
+
+    # q block only
+    qaic_config = dict(enable_blocking=True, num_q_blocks=NUM_Q_BLOCKS)
+    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
+        model_name,
+        continuous_batching=True,
+        transform_params={**transform_params, "qaic_config": qaic_config},
+        export_params=export_params,
+        compile_params=compile_params,
+        generate_params=generate_params,
+        export_compile_only=True,
+    )
+
+    # qkv blocking
+    qaic_config = dict(enable_blocking=True, num_kv_blocks=NUM_KV_BLOCKS, num_q_blocks=NUM_Q_BLOCKS)
+    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
+        model_name,
+        continuous_batching=True,
+        transform_params={**transform_params, "qaic_config": qaic_config},
+        export_params=export_params,
+        compile_params=compile_params,
+        generate_params=generate_params,
+        export_compile_only=True,
+    )
+
+    # head qkv blocking
+    qaic_config = dict(
+        enable_blocking=True,
+        head_block_size=HEAD_BLOCK_SIZE,
+        num_kv_blocks=NUM_KV_BLOCKS,
+        num_q_blocks=NUM_Q_BLOCKS,
+    )
+    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
+        model_name,
+        continuous_batching=True,
+        transform_params={**transform_params, "qaic_config": qaic_config},
+        export_params=export_params,
+        compile_params=compile_params,
+        generate_params=generate_params,
+        export_compile_only=True,
+    )
+
+
+@pytest.mark.llm
+@pytest.mark.qaic
+@pytest.mark.parametrize("model_name", test_models_causal)
+def test_fp16_export_compile_generate_blocking_CB(model_name):
+    """
+    Fp16 + Subfunction + CB + Blocking enabled
+    """
+    if causal_lm_models_dict.get(model_name, None) == model_name and QEFF_TEST_PROFILE == "tiny_model":
+        pytest.skip("Skipping it is not a tiny model and will run in nightly tests.")
+
+    HEAD_BLOCK_SIZE = 8
+    NUM_KV_BLOCKS = 2
+    NUM_Q_BLOCKS = 2
+
+    # head blocking only
+    qaic_config = dict(enable_blocking=True, head_block_size=HEAD_BLOCK_SIZE)
+    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
+        model_name,
+        continuous_batching=True,
+        transform_params={**transform_params, "qaic_config": qaic_config},
+        export_params=export_params,
+        compile_params=compile_params,
+        generate_params=generate_params,
+        export_compile_only=False,
+    )
+
+    # kv blocking only
+    qaic_config = dict(enable_blocking=True, num_kv_blocks=NUM_KV_BLOCKS)
+    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
+        model_name,
+        continuous_batching=True,
+        transform_params={**transform_params, "qaic_config": qaic_config},
+        export_params=export_params,
+        compile_params=compile_params,
+        generate_params=generate_params,
+        export_compile_only=False,
+    )
+
+    # q block only
+    qaic_config = dict(enable_blocking=True, num_q_blocks=NUM_Q_BLOCKS)
+    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
+        model_name,
+        continuous_batching=True,
+        transform_params={**transform_params, "qaic_config": qaic_config},
+        export_params=export_params,
+        compile_params=compile_params,
+        generate_params=generate_params,
+        export_compile_only=False,
+    )
+
+    # qkv blocking
+    qaic_config = dict(enable_blocking=True, num_kv_blocks=NUM_KV_BLOCKS, num_q_blocks=NUM_Q_BLOCKS)
+    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
+        model_name,
+        continuous_batching=True,
+        transform_params={**transform_params, "qaic_config": qaic_config},
+        export_params=export_params,
+        compile_params=compile_params,
+        generate_params=generate_params,
+        export_compile_only=False,
+    )
+
+    # head qkv blocking
+    qaic_config = dict(
+        enable_blocking=True,
+        head_block_size=HEAD_BLOCK_SIZE,
+        num_kv_blocks=NUM_KV_BLOCKS,
+        num_q_blocks=NUM_Q_BLOCKS,
+    )
+    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
+        model_name,
+        continuous_batching=True,
+        transform_params={**transform_params, "qaic_config": qaic_config},
+        export_params=export_params,
+        compile_params=compile_params,
+        generate_params=generate_params,
+        export_compile_only=False,
+    )
