+++ tests/transformers/models/causal_lm_models/test_speculative.py
@@ -1,0 +1,64 @@
+# -----------------------------------------------------------------------------
+#
+# Copyright (c) Qualcomm Technologies, Inc. and/or its subsidiaries.
+# SPDX-License-Identifier: BSD-3-Clause
+#
+# -----------------------------------------------------------------------------
+
+
+import pytest
+
+from QEfficient.utils.constants import Constants
+
+from .check_causal_models import check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100
+from .config import (
+    QEFF_TEST_PROFILE,
+    causal_lm_models_dict,
+    compile_params,
+    export_params,
+    generate_params,
+    test_models_causal,
+    transform_params,
+)
+
+
+@pytest.mark.llm
+@pytest.mark.non_qaic
+@pytest.mark.parametrize("model_name", test_models_causal)
+def test_fp16_export_compile_speculative_cb(model_name):
+    """
+    Fp16 + Subfunction + speculation + CB  (end to end run+ output verification)
+    """
+    if causal_lm_models_dict.get(model_name, None) == model_name and QEFF_TEST_PROFILE == "tiny_model":
+        pytest.skip("Skipping it is not a tiny model and will run in nightly tests.")
+
+    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
+        model_name,
+        continuous_batching=True,
+        transform_params=transform_params,
+        export_params=export_params,
+        compile_params={**compile_params, "num_speculative_tokens": Constants.NUM_SPECULATIVE_TOKENS},
+        generate_params=generate_params,
+        export_compile_only=True,
+    )
+
+
+@pytest.mark.llm
+@pytest.mark.qaic
+@pytest.mark.parametrize("model_name", test_models_causal)
+def test_fp16_export_compile_generate_speculative_cb(model_name):
+    """
+    Fp16 + Subfunction + speculation + CB  (end to end run+ output verification)
+    """
+    if causal_lm_models_dict.get(model_name, None) == model_name and QEFF_TEST_PROFILE == "tiny_model":
+        pytest.skip("Skipping it is not a tiny model and will run in nightly tests.")
+
+    check_causal_lm_pytorch_vs_kv_vs_ort_vs_ai100(
+        model_name,
+        continuous_batching=True,
+        transform_params=transform_params,
+        export_params=export_params,
+        compile_params={**compile_params, "num_speculative_tokens": Constants.NUM_SPECULATIVE_TOKENS},
+        generate_params=generate_params,
+        export_compile_only=True,
+    )
